requirements.txt
kurigram>=2.0.106

aiosqlite>=0.19.0

pytgcrypto>=1.2.5
pyaes>=1.6.1
ujson>=5.7.0
beautifulsoup4>=4.12.2
aiohttp>=3.8.4
aiofiles>=23.1.0
Pillow>=10.3.0
psutil>=5.9.5
humanize>=4.15.0
GitPython>=3.1.30
yt-dlp

pySmartDL>=1.3.4

python-dotenv>=1.0.0
# Media


# AI
# AI
openai>=1.0.0
google-generativeai

# Requests kept for compatibility with plugins until refactored
requests>=2.28.2
motor>=3.3.0
redis>=5.0.0
markdown>=3.4.3

